prevent re-rendering of layout on /about mousemove
